+++ mobile/app/src/main/java/com/bdpsolutions/highfive/services/mediaupload/uploader/ImageUploaderStrategy.kt
@@ -13,7 +13,7 @@ import javax.inject.Inject
 class ImageUploaderStrategy constructor(repoFactory: RepositoryFactory) : IUploaderStrategy(repoFactory) {
 
     override fun uploadFile(path: String,
-                            progressObserver: Observer<Double>,
+                            progressObserver: Observer<Int>,
                             resultObserver: Observer<Result<String>>
     ) {
         (repoFactory.createRepository(RepositoryTypes.IMAGE_REPOSITORY) as ImageRepository)
